fix(app): Handle breaking  entries in relative path to the resource
Previously, the loader was placing the outputs of `.ngux` files in directories next to the dev or prod folders because the relative path to the resource was naïvely used relative to the outputRoot. Now, if the relative resource path begins with a `..` (which implies the next path component is the module name, in a [`generator-mcfly-ng2`](https://github.com/mcfly-io/generator-mcfly-ng2) project) the `..` will be replaced with `external-ux`, placing the files in a separate dir, but still inside the correct dist folder. Subsequent `..`s are discarded.